perf(es/base): Use `bound_names` instead of `find_pat_ids` to improve performance
