+++ crates/swc_typescript/src/fast_dts/util/ast_ext.rs
@@ -6,7 +6,6 @@ use swc_ecma_ast::{BindingIdent, Expr, Lit, MemberProp, ObjectPatProp, Pat, Prop
 pub trait PatExt {
     fn get_type_ann(&self) -> &Option<Box<TsTypeAnn>>;
     fn set_type_ann(&mut self, type_anno: Option<Box<TsTypeAnn>>);
-    fn bound_names<F: FnMut(&BindingIdent)>(&self, f: &mut F);
 }
 
 impl PatExt for Pat {
@@ -37,31 +36,6 @@ impl PatExt for Pat {
             Pat::Rest(rest_pat) => rest_pat.type_ann = type_anno,
             Pat::Object(object_pat) => object_pat.type_ann = type_anno,
             Pat::Assign(_) | Pat::Invalid(_) | Pat::Expr(_) => {}
-        }
-    }
-
-    fn bound_names<F: FnMut(&BindingIdent)>(&self, f: &mut F) {
-        match self {
-            Pat::Ident(binding_ident) => f(binding_ident),
-            Pat::Array(array_pat) => {
-                for pat in array_pat.elems.iter().flatten() {
-                    pat.bound_names(f);
-                }
-            }
-            Pat::Rest(rest_pat) => rest_pat.arg.bound_names(f),
-            Pat::Object(object_pat) => {
-                for pat in &object_pat.props {
-                    match pat {
-                        ObjectPatProp::KeyValue(key_value_pat_prop) => {
-                            key_value_pat_prop.value.bound_names(f)
-                        }
-                        ObjectPatProp::Assign(assign_pat_prop) => f(&assign_pat_prop.key),
-                        ObjectPatProp::Rest(rest_pat) => rest_pat.arg.bound_names(f),
-                    }
-                }
-            }
-            Pat::Assign(assign_pat) => assign_pat.left.bound_names(f),
-            Pat::Invalid(_) | Pat::Expr(_) => todo!(),
         }
     }
 }
